Add javadoc comments and adjust code style
