--- src/main/java/Task.java
@@ -1,22 +1,0 @@
-public class Task {
-    private String description;
-    private boolean done;
-
-    public Task(String description) {
-        this.description = description;
-        this.done = false;
-    }
-
-    public void changeDone(boolean done) {
-        this.done = done;
-    }
-
-    @Override
-    public String toString() {
-        if (this.done) {
-            return "[X] " + this.description;
-        } else {
-            return "[ ] " + this.description;
-        }
-    }
-}
