+++ src/main/java/yapbot/tasks/Deadline.java
@@ -1,6 +1,17 @@
+package yapbot.tasks;
+
+/**
+ * Child class of Task that has an additional deadline field.
+ */
 public class Deadline extends Task {
     private String deadline;
 
+    /**
+     * Returns a Deadline instance.
+     *
+     * @param description Details of the Task.
+     * @param deadline Date/time when this task should be completed by.
+     */
     public Deadline(String description, String deadline) {
         super(description);
         this.deadline = deadline;
